split plugins in a separate directory
