+++ lua/configs/init.lua
@@ -1,0 +1,1 @@
+require 'configs.lazy_init'
