+++ lua/configs/lazy/comment.lua
@@ -1,0 +1,9 @@
+return {
+  'numToStr/Comment.nvim',
+  dependencies = 'JoosepAlviste/nvim-ts-context-commentstring',
+  config = function()
+    require('Comment').setup {
+      pre_hook = require('ts_context_commentstring.integrations.comment_nvim').create_pre_hook(),
+    }
+  end,
+}
