+++ lua/configs/lazy/indent-blankLine.lua
@@ -1,0 +1,7 @@
+return { -- Add indentation guides even on blank lines
+  'lukas-reineke/indent-blankline.nvim',
+  -- Enable `lukas-reineke/indent-blankline.nvim`
+  -- See `:help ibl`
+  main = 'ibl',
+  opts = {},
+}
